+++ src/main/java/com/butterfield/paycheckapp/controller/PaycheckController.java
@@ -3,8 +3,11 @@ package com.butterfield.paycheckapp.controller;
 import com.butterfield.paycheckapp.database.dao.PaycheckDAO;
 import com.butterfield.paycheckapp.database.entity.Paycheck;
 import com.butterfield.paycheckapp.formBean.PaycheckFormBean;
+import com.butterfield.paycheckapp.service.PaycheckService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
@@ -12,16 +15,20 @@ import org.springframework.web.servlet.ModelAndView;
 import javax.validation.Valid;
 import java.util.List;
 
+@Slf4j
 @Controller
 public class PaycheckController {
 
     @Autowired
     private PaycheckDAO paycheckDAO;
 
+    @Autowired
+    private PaycheckService paycheckService = new PaycheckService();
+
     @RequestMapping(value="paycheck/paycheck", method = RequestMethod.GET)
     public ModelAndView emptyPaycheckPage() throws Exception {
         ModelAndView response = new ModelAndView();
-        response.setViewName("paycheck/paycheck");
+        response.setViewName("paycheckInfo");
         return response;
     }
 
@@ -33,28 +40,48 @@ public class PaycheckController {
         System.out.print(paychecks.stream().toList());
         response.addObject("paycheckList", paychecks);
 
-        response.setViewName("paycheckList");
+        response.setViewName("paycheck/paycheckList");
         return response;
     }
+
 
     @RequestMapping(value="paycheck/saveCheck", method = RequestMethod.POST)
     public ModelAndView saveCheck(@Valid PaycheckFormBean form) throws Exception {
         ModelAndView response = new ModelAndView();
 
-        //Adding Check
-        Paycheck paycheck = new Paycheck();
-        paycheck.setPaycheckName(form.getPaycheckName());
-        paycheck.setPaycheckAmount(form.getPaycheckAmount());
-        paycheck.setPaycheckDate(form.getPaycheckDate());
-        paycheck.setDayName(form.getDayName());
-        paycheck.setDaysBetweenCheck(form.getDaysBetweenCheck());
+        Paycheck paycheck = null;
+        log.debug("Incoming form ID: " );
 
+        if (form.getId() == null) {
+            paycheck = new Paycheck();
+            paycheck = paycheckService.saveCheck(form, paycheck);
+            response.setViewName("redirect:../paycheck/list");
+        }
+        else{
+            paycheck = paycheckDAO.findById(form.getId());
+            paycheck = paycheckService.saveCheck(form, paycheck);
+            response.setViewName("redirect:../paycheck/"+ paycheck.getId());
+        }
         paycheckDAO.save(paycheck);
-//        response.addObject(paycheck);
-//        response.setViewName("redirect:../index/" + paycheck.getId());
-        response.setViewName("redirect:../paycheck/list");
-//        indexWithPaycheck();
 
         return response;
     }
+
+    @RequestMapping(value = "/paycheck/{aID}", method = {RequestMethod.GET, RequestMethod.POST})
+    public ModelAndView indexWithPaycheck(@PathVariable("aID") Integer id) throws Exception {
+        ModelAndView response = new ModelAndView();
+
+        Paycheck paycheck = new Paycheck();
+
+        if (paycheckDAO.findById(id) == null) {
+            log.debug("No check!");
+        }else{
+            paycheck = paycheckDAO.findById(id);
+        }
+
+        response.addObject("paycheck", paycheck);
+        response.setViewName("paycheck/paycheckInfo");
+        return response;
+    }
+    
 }
