+++ src/main/resources/templates/index.html
@@ -1,10 +1,9 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head >
-    <div th:replace="~{header :: myHead}"></div>
+<head th:replace="~{header :: myHead}">
 </head>
 <body>
   <h2>Hello from Thyme</h2>
-  <a href="paycheck/list">list check</a>
+  <a href="../paycheck/list">list check</a>
 </body>
 </html>
